Removed code duplication in worker
But keept the setup prepared to easily un-share it for future Play versions.
